+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added `IHaveAzureDevOpsRepository` interface for interacting with Azure DevOps repositories.
   - Added `IPullRequest` interface for managing pull requests on Azure DevOps.
   - Added `IGitFlowWithPullRequest` interface for workflows involving pull request creation on Azure DevOps.
+  - Added `IGitHubFlowWithPullRequest` interface for workflows involving pull request creation on Azure DevOps.
   - Updated dependencies to include Azure DevOps client library.
 
 ## [1.2.1] / 2025-09-14